+++ src/components/Charts/customTooltip.tsx
@@ -5,7 +5,8 @@ import React from 'react'
 import { Box, Divider, Flex, Text, HStack } from '@chakra-ui/react'
 
 // Utils Imports
-import { formatValueToMoney } from '../../../../utils/formatValueToMoney'
+import { formatValueToMoney } from '../../utils/formatValueToMoney'
+
 
 type AreaChartToolTipProps = {
   active?: boolean
